requirements.txt
colorama
tabulate
python-whois
